Move BluetoothState data class to bluetooth package
