+++ app/src/main/java/ecb/manifest/kowalski/obd_scan/ui/viewModels/MainActivityViewModel.kt
@@ -1,14 +1,5 @@
 package ecb.manifest.kowalski.obd_scan.ui.viewModels
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import dagger.hilt.android.lifecycle.HiltViewModel
-import ecb.manifest.kowalski.obd_scan.bluetooth.IBluetoothController
-import ecb.manifest.kowalski.obd_scan.ui.presentation.BluetoothUiState
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.combine
-import kotlinx.coroutines.flow.stateIn
-import javax.inject.Inject
 
 class MainActivityViewModel : ViewModel() { }
